feat: add tests for documentset
